Autonomous Template
Needs to change Names of kConstants for Autonomous to be more understandable for starting positions, doesn't have coding yet for the starting positions (other than for getting off the line for one mode) So needs that - which will require testing due to it being an open loop. The autonomous switch's also need outtake command.